backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.1.1
pydantic==1.10.12
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pymongo==4.4.0
python-dotenv==1.0.0
pycryptodome==3.20.0

celery[redis]==5.3.0
redis==4.6.0
kombu==5.3.0


Pillow
qrcode
fpdf2
